Feat : 지하철 위치 정보를 통한 푸시 알람 구현 완료

+++ src/main/java/com/example/hanium2023/domain/dto/publicapi/arrivalinfo/ArrivalInfoApiResult.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ArrivalInfoApiResult {
     @JsonProperty("barvlDt")
-    private long arrivalTime;
+    private int arrivalTime;
 
     @JsonProperty("updnLine")
     private String direction;
